Add last remaining prop types
