+++ assets/src/block-editor/blocks/amp-brid-player/index.js
@@ -2,19 +2,12 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { InspectorControls } from '@wordpress/block-editor';
-import { Fragment } from '@wordpress/element';
-import {
-	PanelBody,
-	TextControl,
-	Placeholder,
-	ToggleControl,
-} from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import { LayoutControls, MediaPlaceholder } from '../../../components';
+import edit from './edit';
+import save from './save';
 
 export const name = 'amp/amp-brid-player';
 
@@ -74,95 +67,7 @@ export const settings =	{
 		},
 	},
 
-	edit( props ) {
-		const { attributes, setAttributes } = props;
-		const { autoPlay, dataPartner, dataPlayer, dataVideo, dataPlaylist, dataOutstream } = attributes;
-		const ampLayoutOptions = [
-			{ value: 'responsive', label: __( 'Responsive', 'amp' ) },
-			{ value: 'fixed-height', label: __( 'Fixed height', 'amp' ) },
-			{ value: 'fixed', label: __( 'Fixed', 'amp' ) },
-			{ value: 'fill', label: __( 'Fill', 'amp' ) },
-			{ value: 'flex-item', label: __( 'Flex-item', 'amp' ) },
-			{ value: 'nodisplay', label: __( 'No Display', 'amp' ) },
+	edit,
 
-		];
-		let url = false;
-		if ( dataPartner && dataPlayer && ( dataVideo || dataPlaylist || dataOutstream ) ) {
-			url = `http://cdn.brid.tv/live/partners/${ dataPartner }`;
-		}
-		return (
-			<Fragment>
-				<InspectorControls>
-					<PanelBody title={ __( 'Brid Player Settings', 'amp' ) }>
-						<TextControl
-							label={ __( 'Brid.tv partner ID (required)', 'amp' ) }
-							value={ dataPartner }
-							onChange={ ( value ) => ( setAttributes( { dataPartner: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Brid.tv player ID (required)', 'amp' ) }
-							value={ dataPlayer }
-							onChange={ ( value ) => ( setAttributes( { dataPlayer: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Video ID (one of video / playlist / outstream ID is required)', 'amp' ) }
-							value={ dataVideo }
-							onChange={ ( value ) => ( setAttributes( { dataVideo: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Outstream unit ID (one of video / playlist / outstream ID is required)', 'amp' ) }
-							value={ dataOutstream }
-							onChange={ ( value ) => ( setAttributes( { dataOutstream: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Playlist ID (one of video / playlist / outstream ID is required)', 'amp' ) }
-							value={ dataPlaylist }
-							onChange={ ( value ) => ( setAttributes( { dataPlaylist: value } ) ) }
-						/>
-						<ToggleControl
-							label={ __( 'Autoplay', 'amp' ) }
-							checked={ autoPlay }
-							onChange={ () => ( setAttributes( { autoPlay: ! autoPlay } ) ) }
-						/>
-						<LayoutControls { ...props } ampLayoutOptions={ ampLayoutOptions } />
-					</PanelBody>
-				</InspectorControls>
-				{ url && <MediaPlaceholder name={ __( 'Brid Player', 'amp' ) } url={ url } /> }
-				{
-					! url && (
-						<Placeholder label={ __( 'Brid Player', 'amp' ) }>
-							<p>{ __( 'Add required data to use the block.', 'amp' ) }</p>
-						</Placeholder>
-					)
-				}
-			</Fragment>
-		);
-	},
-
-	save( { attributes } ) {
-		const bridProps = {
-			layout: attributes.ampLayout,
-			height: attributes.height,
-			'data-player': attributes.dataPlayer,
-			'data-partner': attributes.dataPartner,
-		};
-		if ( 'fixed-height' !== attributes.ampLayout && attributes.width ) {
-			bridProps.width = attributes.width;
-		}
-		if ( attributes.dataPlaylist ) {
-			bridProps[ 'data-playlist' ] = attributes.dataPlaylist;
-		}
-		if ( attributes.dataVideo ) {
-			bridProps[ 'data-video' ] = attributes.dataVideo;
-		}
-		if ( attributes.dataOutstream ) {
-			bridProps[ 'data-outstream' ] = attributes.dataOutstream;
-		}
-		if ( attributes.autoPlay ) {
-			bridProps.autoplay = attributes.autoPlay;
-		}
-		return (
-			<amp-brid-player { ...bridProps }></amp-brid-player>
-		);
-	},
+	save,
 };
