+++ assets/src/block-editor/blocks/amp-reach-player/index.js
@@ -2,18 +2,12 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { InspectorControls } from '@wordpress/block-editor';
-import { Fragment } from '@wordpress/element';
-import {
-	PanelBody,
-	TextControl,
-	Placeholder,
-} from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import { LayoutControls, MediaPlaceholder } from '../../../components';
+import edit from './edit';
+import save from './save';
 
 export const name = 'amp/amp-reach-player';
 
@@ -53,58 +47,7 @@ export const settings = {
 		},
 	},
 
-	edit( props ) {
-		const { attributes, setAttributes } = props;
-		const { dataEmbedId } = attributes;
-		const ampLayoutOptions = [
-			{ value: 'responsive', label: __( 'Responsive', 'amp' ) },
-			{ value: 'fixed-height', label: __( 'Fixed Height', 'amp' ) },
-			{ value: 'fixed', label: __( 'Fixed', 'amp' ) },
-			{ value: 'fill', label: __( 'Fill', 'amp' ) },
-			{ value: 'flex-item', label: __( 'Flex-item', 'amp' ) },
+	edit,
 
-		];
-		let url = false;
-		if ( dataEmbedId ) {
-			url = 'https://media-cdn.beachfrontreach.com/acct_1/video/';
-		}
-		return (
-			<Fragment>
-				<InspectorControls>
-					<PanelBody title={ __( 'Reach settings', 'amp' ) }>
-						<TextControl
-							label={ __( 'The Reach player embed id (required)', 'amp' ) }
-							value={ dataEmbedId }
-							onChange={ ( value ) => ( setAttributes( { dataEmbedId: value } ) ) }
-						/>
-						<LayoutControls { ...props } ampLayoutOptions={ ampLayoutOptions } />
-					</PanelBody>
-				</InspectorControls>
-				{ url && <MediaPlaceholder name={ __( 'Reach Player', 'amp' ) } url={ url } /> }
-				{
-					! url && (
-						<Placeholder label={ __( 'Reach Player', 'amp' ) }>
-							<p>{ __( 'Add Reach player embed ID to use the block.', 'amp' ) }</p>
-						</Placeholder>
-					)
-				}
-			</Fragment>
-		);
-	},
-
-	save( { attributes } ) {
-		const { dataEmbedId, ampLayout, height, width } = attributes;
-
-		const reachProps = {
-			layout: ampLayout,
-			height,
-			'data-embed-id': dataEmbedId,
-		};
-		if ( 'fixed-height' !== ampLayout && width ) {
-			reachProps.width = width;
-		}
-		return (
-			<amp-reach-player { ...reachProps }></amp-reach-player>
-		);
-	},
+	save,
 };
