+++ assets/src/block-editor/blocks/amp-springboard-player/index.js
@@ -2,19 +2,12 @@
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { InspectorControls } from '@wordpress/block-editor';
-import { Fragment } from '@wordpress/element';
-import {
-	PanelBody,
-	TextControl,
-	Placeholder,
-	SelectControl,
-} from '@wordpress/components';
 
 /**
  * Internal dependencies
  */
-import { LayoutControls, MediaPlaceholder } from '../../../components';
+import edit from './edit';
+import save from './save';
 
 export const name = 'amp/amp-springboard-player';
 
@@ -80,91 +73,7 @@ export const settings = {
 		},
 	},
 
-	edit( props ) {
-		const { attributes, setAttributes } = props;
-		const { dataSiteId, dataPlayerId, dataContentId, dataDomain, dataMode, dataItems } = attributes;
-		const ampLayoutOptions = [
-			{ value: 'responsive', label: __( 'Responsive', 'amp' ) },
-			{ value: 'fixed', label: __( 'Fixed', 'amp' ) },
-			{ value: 'fill', label: __( 'Fill', 'amp' ) },
-			{ value: 'flex-item', label: __( 'Flex-item', 'amp' ) },
+	edit,
 
-		];
-		let url = false;
-		if ( dataSiteId && dataContentId && dataDomain && dataMode && dataItems ) {
-			url = 'https://cms.springboardplatform.com/embed_iframe/';
-		}
-		return (
-			<Fragment>
-				<InspectorControls>
-					<PanelBody title={ __( 'Springboard Player Settings', 'amp' ) }>
-						<TextControl
-							label={ __( 'SprintBoard site ID (required)', 'amp' ) }
-							value={ dataSiteId }
-							onChange={ ( value ) => ( setAttributes( { dataSiteId: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Player content ID (required)', 'amp' ) }
-							value={ dataContentId }
-							onChange={ ( value ) => ( setAttributes( { dataContentId: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Player ID', 'amp' ) }
-							value={ dataPlayerId }
-							onChange={ ( value ) => ( setAttributes( { dataPlayerId: value } ) ) }
-						/>
-						<TextControl
-							label={ __( 'Springboard partner domain', 'amp' ) }
-							value={ dataDomain }
-							onChange={ ( value ) => ( setAttributes( { dataDomain: value } ) ) }
-						/>
-						<SelectControl
-							label={ __( 'Mode (required)', 'amp' ) }
-							value={ dataMode }
-							options={ [
-								{ value: 'video', label: __( 'Video', 'amp' ) },
-								{ value: 'playlist', label: __( 'Playlist', 'amp' ) },
-							] }
-							onChange={ ( value ) => ( setAttributes( { dataMode: value } ) ) }
-						/>
-						<TextControl
-							type="number"
-							label={ __( 'Number of video is playlist (required)', 'amp' ) }
-							value={ dataItems }
-							onChange={ ( value ) => ( setAttributes( { dataItems: value } ) ) }
-						/>
-						<LayoutControls { ...props } ampLayoutOptions={ ampLayoutOptions } />
-					</PanelBody>
-				</InspectorControls>
-				{ url && <MediaPlaceholder name={ __( 'Springboard Player', 'amp' ) } url={ url } /> }
-				{
-					! url && (
-						<Placeholder label={ __( 'Springboard Player', 'amp' ) }>
-							<p>{ __( 'Add required data to use the block.', 'amp' ) }</p>
-						</Placeholder>
-					)
-				}
-			</Fragment>
-		);
-	},
-
-	save( { attributes } ) {
-		const { dataSiteId, dataPlayerId, dataContentId, dataDomain, dataMode, dataItems, ampLayout, height, width } = attributes;
-		const springboardProps = {
-			layout: ampLayout,
-			height,
-			'data-site-id': dataSiteId,
-			'data-mode': dataMode,
-			'data-content-id': dataContentId,
-			'data-player-id': dataPlayerId,
-			'data-domain': dataDomain,
-			'data-items': dataItems,
-		};
-		if ( 'fixed-height' !== ampLayout && width ) {
-			springboardProps.width = attributes.width;
-		}
-		return (
-			<amp-springboard-player { ...springboardProps }></amp-springboard-player>
-		);
-	},
+	save,
 };
