+++ assets/src/components/block-mover/index.js
@@ -1,12 +1,15 @@
 /**
  * This file is mainly copied from the default BlockMover component, there are some small differences.
  * The arrows' labels are changed and are switched. Also, dragging is enabled even if the element is the only block.
+ *
+ * In addition, this copy also contains prop types.
  **/
 
 /**
  * External dependencies
  */
 import classnames from 'classnames';
+import PropTypes from 'prop-types';
 
 /**
  * WordPress dependencies
@@ -67,5 +70,12 @@ export class BlockMover extends Component {
 		);
 	}
 }
+
+BlockMover.propTypes = {
+	isDraggable: PropTypes.bool,
+	onDragStart: PropTypes.func,
+	clientId: PropTypes.string,
+	blockElementId: PropTypes.string,
+};
 
 export default BlockMover;
